add a .gitkeep file to the images/sprites directory
When generating a new project, this directory will be empty by default,
causing git to ignore it. This empty file will ensure git is aware
of this directory.

The gulp file is constructed in such a way that an error will occur if this
directory is not present, so when I gave my friend a copy of the git
repository of my new project, they were not able to run the project
until adding this directory manually.